+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- Added `DataFrame.ja.to_csv` method to write DataFrame to CSV with specified encoding.
+
 ## [0.1.0] - 2024-04-18
 
 ### Added